+++ tests/testthat/test-read-xml.R
@@ -1,5 +1,15 @@
 context("read_xml")
 
+test_that("read_xml errors with an empty document", {
+  expect_error(read_xml(character()), "Document is empty")
+
+  tf <- tempfile()
+  file.create(tf)
+  on.exit(unlink(tf))
+
+  expect_error(read_xml(tf), "Document is empty")
+})
+
 test_that("read_html correctly parses malformed document", {
   lego <- read_html("lego.html.bz2")
   expect_equal(length(xml_find_all(lego, ".//p")), 39)